Implement time-handling logic
Send exactly 1 digest every Monday, or if 1 week has elapsed since the
last digest (in case of an error or if the code isn't triggered on the
preceeding Monday).

Filter query results for papers updated after last digest.  It seems the
arXiv API doesn't support this on their end.

Save the last digest time to an .ini file and only update it if all goes
well.